pyproject.toml
[project]
name = "flask-micro"
version = "0.1.0"
description = ""
authors = [
    {name = "dilozy",email = "denfil228@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "ruff (>=0.12.11,<0.13.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
lint.extend-select = [
    "S", "FBT", "B",
    "A", "COM", "C4",
    "DTZ", "EM", "ISC001",
    "ISC003", "ICN", "PIE",
    "T20", "PYI", "PT",
    "Q", "RSE", "RET",
    "SLF", "SIM", "SLOT",
    "TID", "FLY", "I",
    "N", "PERF", "E",
    "W", "F", "PL",
    "UP", "FURB", "RUF",
    "TRY",
]

lint.ignore = [
    "S101", "RUF012", "SLF001",
    "E501" , "PLR2004"
]